moved some functions out to the lib/colors.js file
